Refactor validatio 생성n
