+++ app/components/user/presentation/PasswordChange.tsx
@@ -1,6 +1,6 @@
 import { useState, useRef, type ChangeEvent } from 'react';
 import { Eye } from 'lucide-react';
-import { isValidPassword } from '@/app/shared/utils/validation';
+import { isValidPassword } from '@/app/shared/consts/validation';
 import { Button } from '@/app/shared/ui/button/Button';
 import { cn } from '@/app/shared/utils/cn';
